+++ packages/nimble-blazor/Examples/Demo.Shared/Pages/ComponentsDemo.razor.cs
@@ -15,6 +15,7 @@ namespace Demo.Shared.Pages
         private NimbleDialog<DialogResult>? _dialog;
         private string? DialogClosedReason { get; set; }
         private NimbleDrawer<DialogResult>? _drawer;
+        private NimbleTable<SimpleTableRecord>? _table;
         private string? DrawerClosedReason { get; set; }
         private string? SelectedRadio { get; set; } = "2";
         private bool BannerOpen { get; set; }
@@ -25,6 +26,12 @@ namespace Demo.Shared.Pages
         public ComponentsDemo()
         {
             AddTableRows(10);
+        }
+
+        protected override async Task OnAfterRenderAsync(bool firstRender)
+        {
+            await _table!.SetDataAsync(TableData);
+            await base.OnAfterRenderAsync(firstRender);
         }
 
         private string DrawerLocationAsString
@@ -68,6 +75,7 @@ namespace Demo.Shared.Pages
 
                 tableData.Add(new SimpleTableRecord(
                     rowCountString,
+                    tableData.Count >= 4 ? (tableData.Count % 4).ToString(CultureInfo.CurrentCulture) : null,
                     $"new string {rowCountString}",
                     $"bar {rowCountString}",
                     "/",
@@ -87,6 +95,7 @@ namespace Demo.Shared.Pages
     {
         public SimpleTableRecord(
             string id,
+            string? parentId,
             string stringValue1,
             string stringValue2,
             string? href,
@@ -98,6 +107,7 @@ namespace Demo.Shared.Pages
             double duration)
         {
             Id = id;
+            ParentId = parentId;
             StringValue1 = stringValue1;
             StringValue2 = stringValue2;
             Href = href;
@@ -110,6 +120,7 @@ namespace Demo.Shared.Pages
         }
 
         public string Id { get; }
+        public string? ParentId { get; }
         public string StringValue1 { get; }
         public string StringValue2 { get; }
         public string? Href { get; }